requirements.txt
Flask
azure-servicebus
azure-storage-blob
pylint
autopep8